Fix a required property test.
The test ensures an error is generated when a required field is added and subsequently removed. The test removed the
field by modifying a string property to be an empty string, which is still a valid string. This fixes this test by
removing th property by setting it to undefined.